+++ CHANGELOG.md
@@ -31,6 +31,11 @@
   - remove `--compressed` from the CLI. Snapshots are now always compressed.
   - Remove snapshot ops - snapshots fetched to the current directory by default.
 
+### Fixed
+
+- [#2967](https://github.com/ChainSafe/forest/issues/2967): Fix http-client
+  concurrency issues caused by fetching root certificates multiple times.
+
 ### Added
 
 - [#2706](https://github.com/ChainSafe/forest/issues/2706): implement